Time: 134 ms (20.56%), Space: 96.7 MB (22.05%) - LeetHub

+++ 0896-monotonic-array/0896-monotonic-array.cpp
@@ -5,7 +5,7 @@ public:
 
         int direction = 0;  // 0 means unknown, 1 means increasing, -1 means decreasing
 
-        for (size_t i = 1; i < nums.size(); i++) {
+        for (int i = 1; i < nums.size(); i++) {
             if (nums[i] > nums[i-1]) {  // increasing
                 if (direction == 0) direction = 1;
                 else if (direction == -1) return false;
